[Misc] Separate out `utils.counter` and move `utils.Device` to engine (#27588)
